--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "python.pythonPath": "env\\Scripts\\python.exe",
-    "python.linting.pylintEnabled": true,
-    "python.linting.enabled": true
-}
